Fix EnemyController bug
The controller subscribed to events for all objects of GameContext._interactables list, including the player. This caused an error when the player took damage, as the EnemyController was intercepting the subscription to player events that should have been called from BackState class. Now the controller subscribes to events for only objects with a tag "Enemy".